Implement `is` and `isnt`
These are much harder than in Perl where you rely on values just
stringifying somehow. Here, I opted for libtap++'s template that
makes equality parametric as well and defaults to std::equal_to.

BUT I went a bit further in digging up some template voodoo which
tests if the supplied values have an overload of operator<< that
allows me to stringify them. If yes, and the `is` / `isnt` test
failed, I print the differing values, but if not, I let the code
compile without those extra diagnostics.