Enable extra warnings
And silence some unused arguments warnings. Due to an already-but-
only-recently fixed bug in g++ [1] we can't mark the first argument
of a constructor unused, unless the tag goes *after* the name, so
we do that throughout. Not sure if that is standards-conforming,
but clang++ accepts it as well as g++.

[1] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=81429