Support `like` and `unlike`
While `is` and `isnt` compare two ("data") objects for a match,
`like` and `unlike` take one ("data") object and another "predicate"
object and sees if the object matches the predicate.

There are candidates that turn an std::string predicate into an
std::regex and use regex matching as the actual predicate.

This doesn't quite match Raku's behavior, but it seems like a useful
and easily memorizable variant.