Remove std::shared_ptr from public API
I hear that it is bad form to expose smart pointers in the public API
and it didn't feel good to require std::enable_shared_from_this for
making subtests of a Context.

Now the lifetime of parents of subtests must be managed by the user.
The Convenience API continues to do this with std::shared_ptr, but
their usage is now well hidden, and this is enough for me.